feat(security): TASK-003 PC4安全审查基础设施 + 对抗性代码审查
P0修复:
- 移除database.py/config.py硬编码凭据,改用环境变量
- 创建JWT认证模块(security.py) + User模型 + /auth/login端点
- 13个API端点全部添加get_current_user依赖注入

P1修复:
- bidding/projects create强制owner_id/manager_id为当前用户
- update/delete添加越权检查(owner或admin可操作)
- search参数LIKE通配符转义

P2修复:
- organizations update添加祖先链循环引用检测
- 错误信息消除内部状态码泄露

测试:
- 98个pytest测试全部通过(SQLite内存库)
- 覆盖: 模型/401认证/CRUD/分页过滤/越权检查/日历看板树

+++ backend/app/models/bidding.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from sqlalchemy import Column, BigInteger, String, Numeric, SmallInteger, Text, DateTime
+from sqlalchemy import Column, BigInteger, String, Numeric, SmallInteger, Text, DateTime, Integer
 from sqlalchemy.orm import validates
 from app.core.database import Base
 
@@ -15,7 +15,7 @@ VALID_BID_TRANSITIONS = {
 
 class Bidding(Base):
     __tablename__ = "bidding"
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     title = Column(String(200), nullable=False)
     project_name = Column(String(200))
     bid_amount = Column(Numeric(15, 2))
